changed to Adventure's sound object
