+++ common/src/main/java/net/draycia/carbon/common/command/commands/ContinueCommand.java
@@ -132,7 +132,7 @@ public class ContinueCommand extends CarbonCommand {
                 this.carbonMessages.whisperRecipient(new SourcedAudience(sender, recipient), senderName, recipientName, message);
                 this.carbonMessages.whisperConsoleLog(this.carbonChat.server().console(), senderName, recipientName, message);
 
-                recipient.playSound(this.configFactory.primaryConfig().messageSounds().sound());
+                recipient.playSound(this.configFactory.primaryConfig().messageSounds());
                 sender.lastWhisperTarget(recipient.uuid());
                 sender.whisperReplyTarget(recipient.uuid());
                 recipient.whisperReplyTarget(sender.uuid());
