+++ common/src/main/java/net/draycia/carbon/common/command/commands/WhisperCommand.java
@@ -126,7 +126,7 @@ public class WhisperCommand extends CarbonCommand {
                 this.carbonMessages.whisperRecipient(new SourcedAudience(sender, recipient), senderName, recipientName, message);
                 this.carbonMessages.whisperConsoleLog(this.carbonChat.server().console(), senderName, recipientName, message);
 
-                recipient.playSound(this.configFactory.primaryConfig().messageSounds().sound());
+                recipient.playSound(this.configFactory.primaryConfig().messageSounds());
                 sender.lastWhisperTarget(recipient.uuid());
                 sender.whisperReplyTarget(recipient.uuid());
                 recipient.whisperReplyTarget(sender.uuid());
