+++ common/src/main/java/net/draycia/carbon/common/config/PingSettings.java
@@ -19,13 +19,15 @@
  */
 package net.draycia.carbon.common.config;
 
+import net.kyori.adventure.key.Key;
+import net.kyori.adventure.sound.Sound;
 import net.kyori.adventure.text.format.NamedTextColor;
 import net.kyori.adventure.text.format.TextColor;
 import org.spongepowered.configurate.objectmapping.ConfigSerializable;
 import org.spongepowered.configurate.objectmapping.meta.Comment;
 
 @ConfigSerializable
-public class PingSettings extends SoundSettings {
+public class PingSettings {
 
     @Comment("The color your name will be when another player mentions you.")
     private TextColor highlightTextColor = NamedTextColor.YELLOW;
@@ -33,6 +35,10 @@ public class PingSettings extends SoundSettings {
     private String prefix = "@";
 
     private boolean playSound = false;
+    private Key name = Key.key("block.anvil.use");
+    private Sound.Source source = Sound.Source.MASTER;
+    private float volume = 1.0f; // 0.0 -> infinity
+    private float pitch = 1.0f; // 0.0 -> 2.0
 
     public TextColor highlightTextColor() {
         return this.highlightTextColor;
@@ -42,8 +48,28 @@ public class PingSettings extends SoundSettings {
         return this.playSound;
     }
 
+    public Key name() {
+        return this.name;
+    }
+
     public String prefix() {
         return this.prefix;
+    }
+
+    public Sound.Source source() {
+        return this.source;
+    }
+
+    public float volume() {
+        return this.volume;
+    }
+
+    public float pitch() {
+        return this.pitch;
+    }
+
+    public Sound sound() {
+        return Sound.sound(this.name, this.source, this.volume, this.pitch);
     }
 
 }
