+++ common/src/main/java/net/draycia/carbon/common/config/PrimaryConfig.java
@@ -23,6 +23,7 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import net.kyori.adventure.key.Key;
+import net.kyori.adventure.sound.Sound;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.checkerframework.framework.qual.DefaultQualifier;
@@ -70,7 +71,12 @@ public class PrimaryConfig {
     private PingSettings pingSettings = new PingSettings();
 
     @Comment("Various sound settings for messages.")
-    private SoundSettings messageSounds = new SoundSettings();
+    private Sound messageSounds = Sound.sound(
+        Key.key("entity.experience_orb.pickup"),
+        Sound.Source.MASTER,
+        1.0F,
+        1.0F
+    );
 
     private MessagingSettings messagingSettings = new MessagingSettings();
 
@@ -114,7 +120,7 @@ public class PrimaryConfig {
         return this.messagingSettings;
     }
 
-    public SoundSettings messageSounds() {
+    public Sound messageSounds() {
         return this.messageSounds;
     }
